+++ CHANGELOG.rst
@@ -2,6 +2,8 @@ Changelog
 ---------
 
 v1.28 (??? 2023)
+    * Do not set color for missing data; only existed for sron palettes.
+    * Add --dpi as a standard option in `mcPlot`.
     * Use mpl.colormaps[name] instead of mpl.colormaps.get_cmap(name)
       to replace mpl.cm.get_cmap(name) to work with matplotlib < v3.6.